Flyttet det som bare brukes av MicrosoftGraphClient inn i MicrosoftGraphClient
