+++ src/main/kotlin/no/nav/klage/oppgave/clients/MicrosoftGraphClient.kt
@@ -3,6 +3,8 @@ package no.nav.klage.oppgave.clients
 import no.nav.klage.oppgave.domain.MicrosoftGraphIdentResponse
 import no.nav.klage.oppgave.domain.MicrosoftGraphNameResponse
 import no.nav.klage.oppgave.util.getLogger
+import no.nav.security.token.support.client.core.oauth2.OAuth2AccessTokenService
+import no.nav.security.token.support.client.spring.ClientConfigurationProperties
 import org.springframework.retry.annotation.Retryable
 import org.springframework.stereotype.Component
 import org.springframework.web.reactive.function.client.WebClient
@@ -10,7 +12,11 @@ import org.springframework.web.reactive.function.client.bodyToMono
 import kotlin.system.measureTimeMillis
 
 @Component
-class MicrosoftGraphClient(private val microsoftGraphWebClient: WebClient) {
+class MicrosoftGraphClient(
+    private val microsoftGraphWebClient: WebClient,
+    private val clientConfigurationProperties: ClientConfigurationProperties,
+    private val oAuth2AccessTokenService: OAuth2AccessTokenService
+) {
 
     companion object {
         @Suppress("JAVA_CLASS_ON_COMPANION")
@@ -18,8 +24,9 @@ class MicrosoftGraphClient(private val microsoftGraphWebClient: WebClient) {
     }
 
     @Retryable
-    fun getNavIdent(accessToken: String): String {
+    fun getNavIdentForAuthenticatedUser(): String {
         logger.debug("Fetching navIdent from Microsoft Graph")
+        val saksbehandlerAccessToken = getSaksbehandlerTokenWithGraphScope()
 
         return microsoftGraphWebClient.get()
             .uri { uriBuilder ->
@@ -27,7 +34,7 @@ class MicrosoftGraphClient(private val microsoftGraphWebClient: WebClient) {
                     .path("/me")
                     .queryParam("\$select", "onPremisesSamAccountName")
                     .build()
-            }.header("Authorization", "Bearer $accessToken")
+            }.header("Authorization", "Bearer $saksbehandlerAccessToken")
 
             .retrieve()
             .bodyToMono<MicrosoftGraphIdentResponse>()
@@ -35,7 +42,8 @@ class MicrosoftGraphClient(private val microsoftGraphWebClient: WebClient) {
     }
 
     @Retryable
-    fun getDisplayNames(idents: List<String>, accessToken: String): Map<String, String> {
+    fun getDisplayNames(idents: List<String>): Map<String, String> {
+        val appAccessToken = getAppTokenWithGraphScope()
         return try {
             val response = microsoftGraphWebClient.get()
                 .uri { uriBuilder ->
@@ -47,7 +55,7 @@ class MicrosoftGraphClient(private val microsoftGraphWebClient: WebClient) {
                         )
                         .queryParam("\$select", "onPremisesSamAccountName,displayName")
                         .build()
-                }.header("Authorization", "Bearer $accessToken")
+                }.header("Authorization", "Bearer $appAccessToken")
                 .retrieve()
                 .bodyToMono<MicrosoftGraphNameResponse>()
                 .block()
@@ -63,5 +71,17 @@ class MicrosoftGraphClient(private val microsoftGraphWebClient: WebClient) {
             logger.warn("Could not fetch displayname for idents: $idents", e)
             emptyMap()
         }
+    }
+
+    private fun getSaksbehandlerTokenWithGraphScope(): String {
+        val clientProperties = clientConfigurationProperties.registration["onbehalfof"]
+        val response = oAuth2AccessTokenService.getAccessToken(clientProperties)
+        return response.accessToken
+    }
+
+    private fun getAppTokenWithGraphScope(): String {
+        val clientProperties = clientConfigurationProperties.registration["app"]
+        val response = oAuth2AccessTokenService.getAccessToken(clientProperties)
+        return response.accessToken
     }
 }
